Move get_trigger and _never_fire_trigger to util
